+++ api/config/model.yaml
@@ -1,8 +1,4 @@
-save_dir: ./ckpts/v22_bigmodel_mid_newdropout
-data_folder: null
-data_name: flickr8k_1_cap_per_img_1_min_word_freq
 checkpoint_file: ./ckpts/BEST_checkpoint_flickr8k_1_cap_per_img_1_min_word_freq.pth
-word_embedding_weight: null
 word_map_file: ./ckpts/WORDMAP_flickr8k_1_cap_per_img_1_min_word_freq.json
 attention_dim: 512
 emb_dim: 512
@@ -14,21 +10,4 @@ regularized_decoder:
     weight_p: 0.4
     input_p: 0.3
     output_p: 0.3
-device: cuda
-epochs: 30
-workers: 1
-batch_size: 128
-optimizer: ranger
 fine_tune_encoder: false
-encoder_lr: 0.0001
-decoder_lr: 0.004
-grad_clip: 5.0
-alpha_c: 1.0
-confidence_c: null
-best_bleu4: 0.0
-print_freq: 50
-tolerance_epoch: 8
-adjust_epoch: 2
-adjust_step: 0.6
-decay_epoch: 2
-decay_step: 0.8
